perf: cache /AS/{asn} responses at startup
/AS/{asn} previously ran a full linear scan of both ASN mmdb tries on
every request. Mirror the existing COUNTRY_CACHE pattern: build an
ASN_CACHE keyed by ASN number once at startup, and reduce the handler
to a HashMap lookup.

This also fixes a pre-existing bug where ASNs with only IPv6
announcements (no IPv4 presence) incorrectly returned 400 — the old
scan only ever set the returned ASN metadata inside the IPv4 loop.
